Implemented conditional compiling using gcc preprocessor. To prevent a line from being compiled without octagon write: #ifdef INCLUDE_OCT ... source ... #endif
The default compilation is without octagon nor debugging
To compile with debug flags run: make debug=1
To compile with the octagon library run: make oct=1
They can be combined as well.

Caveats:
ocamlc is responsible for calling the preprocessor and it stores the result in a temp file which then passes to the compiler.
The compiler will report errors with the filename of the temp file, although the name of the original file can be read at the end of the previous line in the compiler output.